Fixed Garrett's tech pad AGAIN
For some reason the little gray details with the teleporter special keep flipping inwards so the player can never escape off the pad once they go on, go back on the other teleporter and back because the little lines are facing inwards, not outwards like the remaining lines are. CSG related?